# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ecommerce API",
        "description": "API documentation for the ecommerce application.",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/get-all-products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Retrieve all products",
                "operationId": "1344e57ba1a30117d45f9d8f160ce6dd",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/get-product-by-id": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Retrieve a product by ID",
                "operationId": "fab496ac6ae8c4a592e6e39db8851edb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
        "/api/create-products": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product",
                "operationId": "6972bef5011d78a9084034310202b057",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Smartphone"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 299.99
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A high-end smartphone"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
        "/api/update-product-by-id": {
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update a product",
                "operationId": "c70d2619174cdabf0de88a2c20375379",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Smartphone Pro"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 399.99
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "An upgraded smartphone"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
        "/api/delete-product-by-id": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Soft delete a product",
                "operationId": "8be17d4dab6b21fb5e1eff02c129571b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product deleted successfully"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Product": {
                "required": [
                    "id",
                    "name",
                    "price",
                    "description"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Smartphone"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 299.99
                    },
                    "description": {
                        "type": "string",
                        "example": "A high-end smartphone"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-01T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-01T12:34:56Z"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Products",
            "description": "Products"
        }
    ]
}